Rewrite tests to use test vector.
