+++ lib/test.js
@@ -1,5 +1,7 @@
 'use strict'
 
+const contractsModule = require('./contractManager')
+const DevPipe = require('./pipes/dev_pipe')
 const erisContracts = require('..')
 const I = require('iteray')
 const Promise = require('bluebird')
@@ -7,23 +9,6 @@ const R = require('ramda')
 const Solidity = require('solc')
 const stream = require('stream')
 const test = require('eris-db/lib/test')
-
-const observer = (asyncIterable) => R.pipe(
-  I.map((event) => JSON.stringify(event, null, 2) + '\n\n'),
-  I.to(stream.Readable)
-)(asyncIterable).pipe(process.stderr)
-
-const newContractManager = (name, options) =>
-  Promise.all([
-    test.newBlockchain(name, options),
-    test.privateValidator(name)
-  ]).spread((url, validator) =>
-    erisContracts.newContractManagerDev(url, {
-      address: validator.address,
-      pubKey: validator.pub_key,
-      privKey: validator.priv_key
-    }, {observer: process.env.DEBUG ? observer : I.sink})
-  )
 
 const compile = (contractManager, source, name) => {
   const compiled = Solidity.compile(source, 1).contracts[name]
@@ -35,8 +20,36 @@ const compile = (contractManager, source, name) => {
   )
 }
 
+const Vector = () => {
+  let manager
+  const vector = test.Vector()
+
+  return Object.assign(Object.create(vector), {
+    before: (dirname, options, callback) =>
+      vector.before(dirname, options, function ({db, validator}) {
+        const accounts = {
+          address: validator.address,
+          pubKey: validator.pub_key,
+          privKey: validator.priv_key
+        }
+
+        manager = contractsModule.newContractManager(new DevPipe(db, accounts))
+
+        if (callback) {
+          return callback.call(this, manager)
+        }
+      }),
+
+    it: (callback) =>
+      vector.it(
+        function () {
+          return callback.call(this, manager)
+        }
+      )
+  })
+}
+
 module.exports = {
-  newContractManager,
   compile,
-  observer
+  Vector
 }
