+++ test/#21/index.js
@@ -4,12 +4,12 @@ const assert = require('assert')
 const Promise = require('bluebird')
 const test = require('../../lib/test')
 
+const vector = test.Vector()
+
 describe('issue #21', function () {
   let contract
 
-  before(function () {
-    this.timeout(60 * 1000)
-
+  before(vector.before(__dirname, {protocol: 'http:'}, function (manager) {
     const source = `
       contract c {
         function getBytes() constant returns (byte[10]){
@@ -30,15 +30,14 @@ describe('issue #21', function () {
       }
     `
 
-    return test.newContractManager('blockchain', {protocol: 'http:'})
-      .then((manager) =>
-        test.compile(manager, source, 'c').then((compiledContract) => {
-          contract = compiledContract
-        })
-      )
-  })
+    return test.compile(manager, source, 'c').then((compiledContract) => {
+      contract = compiledContract
+    })
+  }))
 
-  it('gets the static byte array decoded properly', function () {
+  after(vector.after())
+
+  it('gets the static byte array decoded properly', vector.it(function () {
     return Promise.fromCallback((callback) => contract.getBytes(callback))
       .then((bytes) => {
         assert.deepEqual(
@@ -46,12 +45,14 @@ describe('issue #21', function () {
           ['68', '65', '6C', '6C', '6F', '00', '00', '00', '00', 'FF']
         )
       })
-  })
+  }))
 
-  it('returns multiple values correctly from a function', function () {
-    return Promise.fromCallback((callback) => contract.deeper(callback))
-      .then((values) => {
-        assert.equal(Number(values[1]), 42)
-      })
-  })
+  it('returns multiple values correctly from a function',
+    vector.it(function () {
+      return Promise.fromCallback((callback) => contract.deeper(callback))
+        .then((values) => {
+          assert.equal(Number(values[1]), 42)
+        })
+    })
+  )
 })
