+++ test/#47/vector.json
@@ -1,0 +1,119 @@
+{
+  "it": {
+    "#47": [
+      {
+        "input": {
+          "done": false,
+          "value": {
+            "method": "erisdb.transactAndHold",
+            "params": {
+              "priv_key": "86F527B157FF098AC6BEDC38B0478A3099E03F4E55D46429DD365A359706557916137A2137316939F502F2BE428B8A9802EA9A0A0C4A960580D128EDC24E5B4A",
+              "address": "",
+              "data": "60a060405260086060527f2020506965746572000000000000000000000000000000000000000000000000608052600080548180527f202050696574657200000000000000000000000000000000000000000000001082556100b3907f290decd9548b62a8d60345a988386fc84ba6bc95484008f6362f93160ef3e563602060026001841615610100026000190190931692909204601f01919091048101905b80821115610145576000815560010161009f565b505060408051808201909152600681527f50696574657200000000000000000000000000000000000000000000000000006020918201908152600180546000829052915160ff1916600c1781559161014991600281851615610100026000190190911604601f01047fb10e2d527612073b26eecdfd717e6a320cf44b4afac2b0732d9fcbe2b7fa0cf69081019061009f565b5090565b505061018d806101596000396000f3606060405260e060020a60003504638b1d49388114610026578063ada94c0714610089575b005b6100ea600060608181528154602060026001831615610100026000190190921691909104601f810182900490910260a0908101604052608082815292939190828280156101835780601f1061015857610100808354040283529160200191610183565b6100ea600060609081526001805460a06020601f6002600019610100868816150201909416939093049283018190040281016040526080828152929190828280156101835780601f1061015857610100808354040283529160200191610183565b60405180806020018281038252838181518152602001915080519060200190808383829060006004602084601f0104600f02600301f150905090810190601f16801561014a5780820380516001836020036101000a031916815260200191505b509250505060405180910390f35b820191906000526020600020905b81548152906001019060200180831161016657829003601f168201915b505050505090509056",
+              "gas_limit": 1000000,
+              "fee": 0
+            },
+            "jsonrpc": "2.0",
+            "id": "1"
+          }
+        }
+      },
+      {
+        "output": {
+          "done": false,
+          "value": {
+            "result": {
+              "call_data": {
+                "caller": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+                "callee": "CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+                "data": "60A060405260086060527F2020506965746572000000000000000000000000000000000000000000000000608052600080548180527F202050696574657200000000000000000000000000000000000000000000001082556100B3907F290DECD9548B62A8D60345A988386FC84BA6BC95484008F6362F93160EF3E563602060026001841615610100026000190190931692909204601F01919091048101905B80821115610145576000815560010161009F565B505060408051808201909152600681527F50696574657200000000000000000000000000000000000000000000000000006020918201908152600180546000829052915160FF1916600C1781559161014991600281851615610100026000190190911604601F01047FB10E2D527612073B26EECDFD717E6A320CF44B4AFAC2B0732D9FCBE2B7FA0CF69081019061009F565B5090565B505061018D806101596000396000F3606060405260E060020A60003504638B1D49388114610026578063ADA94C0714610089575B005B6100EA600060608181528154602060026001831615610100026000190190921691909104601F810182900490910260A0908101604052608082815292939190828280156101835780601F1061015857610100808354040283529160200191610183565B6100EA600060609081526001805460A06020601F6002600019610100868816150201909416939093049283018190040281016040526080828152929190828280156101835780601F1061015857610100808354040283529160200191610183565B60405180806020018281038252838181518152602001915080519060200190808383829060006004602084601F0104600F02600301F150905090810190601F16801561014A5780820380516001836020036101000A031916815260200191505B509250505060405180910390F35B820191906000526020600020905B81548152906001019060200180831161016657829003601F168201915B505050505090509056",
+                "value": 1,
+                "gas": 999765
+              },
+              "origin": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+              "tx_id": "D981AA38F4D62E4656DBF90031C138EA86118065",
+              "return": "606060405260E060020A60003504638B1D49388114610026578063ADA94C0714610089575B005B6100EA600060608181528154602060026001831615610100026000190190921691909104601F810182900490910260A0908101604052608082815292939190828280156101835780601F1061015857610100808354040283529160200191610183565B6100EA600060609081526001805460A06020601F6002600019610100868816150201909416939093049283018190040281016040526080828152929190828280156101835780601F1061015857610100808354040283529160200191610183565B60405180806020018281038252838181518152602001915080519060200190808383829060006004602084601F0104600F02600301F150905090810190601F16801561014A5780820380516001836020036101000A031916815260200191505B509250505060405180910390F35B820191906000526020600020905B81548152906001019060200180831161016657829003601F168201915B505050505090509056",
+              "exception": ""
+            },
+            "error": null,
+            "id": "1",
+            "jsonrpc": "2.0"
+          }
+        }
+      },
+      {
+        "input": {
+          "done": false,
+          "value": {
+            "method": "erisdb.call",
+            "params": {
+              "from": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+              "address": "CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+              "data": "8b1d4938"
+            },
+            "jsonrpc": "2.0",
+            "id": "2"
+          }
+        }
+      },
+      {
+        "output": {
+          "done": false,
+          "value": {
+            "result": {
+              "return": "000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000082020506965746572000000000000000000000000000000000000000000000000",
+              "gas_used": 327
+            },
+            "error": null,
+            "id": "2",
+            "jsonrpc": "2.0"
+          }
+        }
+      },
+      {
+        "input": {
+          "done": false,
+          "value": {
+            "method": "erisdb.call",
+            "params": {
+              "from": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+              "address": "CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+              "data": "ada94c07"
+            },
+            "jsonrpc": "2.0",
+            "id": "3"
+          }
+        }
+      },
+      {
+        "output": {
+          "done": false,
+          "value": {
+            "result": {
+              "return": "000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000065069657465720000000000000000000000000000000000000000000000000000",
+              "gas_used": 334
+            },
+            "error": null,
+            "id": "3",
+            "jsonrpc": "2.0"
+          }
+        }
+      }
+    ]
+  },
+  "validator": {
+    "address": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+    "pub_key": [
+      1,
+      "16137A2137316939F502F2BE428B8A9802EA9A0A0C4A960580D128EDC24E5B4A"
+    ],
+    "priv_key": [
+      1,
+      "86F527B157FF098AC6BEDC38B0478A3099E03F4E55D46429DD365A359706557916137A2137316939F502F2BE428B8A9802EA9A0A0C4A960580D128EDC24E5B4A"
+    ],
+    "last_height": 0,
+    "last_round": 0,
+    "last_step": 0
+  },
+  "before": []
+}
