+++ test/event/vector.json
@@ -1,0 +1,210 @@
+{
+  "it": {
+    "listens to an event from a contract": [
+      {
+        "input": {
+          "done": false,
+          "value": {
+            "method": "erisdb.transactAndHold",
+            "params": {
+              "priv_key": "86F527B157FF098AC6BEDC38B0478A3099E03F4E55D46429DD365A359706557916137A2137316939F502F2BE428B8A9802EA9A0A0C4A960580D128EDC24E5B4A",
+              "address": "",
+              "data": "606060405260608060106000396000f3606060405260e060020a60003504633bdab8bf8114601a575b005b601873ffffffffffffffffffffffffffffffffffffffff331660609081527f50d7c806d0f7913f321946784dee176a42aa55b5dd83371fc57dcedf659085e090602090a156",
+              "gas_limit": 1000000,
+              "fee": 0
+            },
+            "jsonrpc": "2.0",
+            "id": "1"
+          }
+        }
+      },
+      {
+        "output": {
+          "done": false,
+          "value": {
+            "result": {
+              "call_data": {
+                "caller": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+                "callee": "CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+                "data": "606060405260608060106000396000F3606060405260E060020A60003504633BDAB8BF8114601A575B005B601873FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF331660609081527F50D7C806D0F7913F321946784DEE176A42AA55B5DD83371FC57DCEDF659085E090602090A156",
+                "value": 1,
+                "gas": 999985
+              },
+              "origin": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+              "tx_id": "39AF9B36CC22E9105B74BA338829B95533764375",
+              "return": "606060405260E060020A60003504633BDAB8BF8114601A575B005B601873FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF331660609081527F50D7C806D0F7913F321946784DEE176A42AA55B5DD83371FC57DCEDF659085E090602090A156",
+              "exception": ""
+            },
+            "error": null,
+            "id": "1",
+            "jsonrpc": "2.0"
+          }
+        }
+      },
+      {
+        "input": {
+          "done": false,
+          "value": {
+            "method": "erisdb.eventSubscribe",
+            "params": {
+              "event_id": "Log/CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7"
+            },
+            "jsonrpc": "2.0",
+            "id": "2"
+          }
+        }
+      },
+      {
+        "output": {
+          "done": false,
+          "value": {
+            "result": {
+              "sub_id": "D47F81F822B9CECB9F6452538C46F91C4913C42EB2464DB5EED70E9F77FFB6AF"
+            },
+            "error": null,
+            "id": "2",
+            "jsonrpc": "2.0"
+          }
+        }
+      },
+      {
+        "input": {
+          "done": false,
+          "value": {
+            "method": "erisdb.transactAndHold",
+            "params": {
+              "priv_key": "86F527B157FF098AC6BEDC38B0478A3099E03F4E55D46429DD365A359706557916137A2137316939F502F2BE428B8A9802EA9A0A0C4A960580D128EDC24E5B4A",
+              "address": "CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+              "data": "3bdab8bf",
+              "gas_limit": 1000000,
+              "fee": 0
+            },
+            "jsonrpc": "2.0",
+            "id": "3"
+          }
+        }
+      },
+      {
+        "output": {
+          "done": false,
+          "value": {
+            "result": {
+              "call_data": {
+                "caller": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+                "callee": "CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+                "data": "3BDAB8BF",
+                "value": 1,
+                "gas": 999956
+              },
+              "origin": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+              "tx_id": "C185A9C1EC05D241756FCC2409AF4020E3D27853",
+              "return": "",
+              "exception": ""
+            },
+            "error": null,
+            "id": "3",
+            "jsonrpc": "2.0"
+          }
+        }
+      },
+      {
+        "input": {
+          "done": false,
+          "value": {
+            "method": "erisdb.transactAndHold",
+            "params": {
+              "priv_key": "86F527B157FF098AC6BEDC38B0478A3099E03F4E55D46429DD365A359706557916137A2137316939F502F2BE428B8A9802EA9A0A0C4A960580D128EDC24E5B4A",
+              "address": "CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+              "data": "3bdab8bf",
+              "gas_limit": 1000000,
+              "fee": 0
+            },
+            "jsonrpc": "2.0",
+            "id": "4"
+          }
+        }
+      },
+      {
+        "output": {
+          "done": false,
+          "value": {
+            "result": {
+              "call_data": {
+                "caller": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+                "callee": "CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+                "data": "3BDAB8BF",
+                "value": 1,
+                "gas": 999956
+              },
+              "origin": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+              "tx_id": "88B8FD133CC8108023E8DE321FA3EF31E391F202",
+              "return": "",
+              "exception": ""
+            },
+            "error": null,
+            "id": "4",
+            "jsonrpc": "2.0"
+          }
+        }
+      },
+      {
+        "input": {
+          "done": false,
+          "value": {
+            "method": "erisdb.eventPoll",
+            "params": {
+              "sub_id": "D47F81F822B9CECB9F6452538C46F91C4913C42EB2464DB5EED70E9F77FFB6AF"
+            },
+            "jsonrpc": "2.0",
+            "id": "5"
+          }
+        }
+      },
+      {
+        "output": {
+          "done": false,
+          "value": {
+            "result": {
+              "events": [
+                {
+                  "address": "000000000000000000000000CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+                  "topics": [
+                    "50D7C806D0F7913F321946784DEE176A42AA55B5DD83371FC57DCEDF659085E0"
+                  ],
+                  "data": "000000000000000000000000E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+                  "height": 3
+                },
+                {
+                  "address": "000000000000000000000000CCB6ECAA61B9DF029DE3FB1F1614793D6CE393B7",
+                  "topics": [
+                    "50D7C806D0F7913F321946784DEE176A42AA55B5DD83371FC57DCEDF659085E0"
+                  ],
+                  "data": "000000000000000000000000E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+                  "height": 4
+                }
+              ]
+            },
+            "error": null,
+            "id": "5",
+            "jsonrpc": "2.0"
+          }
+        }
+      }
+    ]
+  },
+  "validator": {
+    "address": "E9978233F1DF5CBD52269B933D42E6203F2D5D71",
+    "pub_key": [
+      1,
+      "16137A2137316939F502F2BE428B8A9802EA9A0A0C4A960580D128EDC24E5B4A"
+    ],
+    "priv_key": [
+      1,
+      "86F527B157FF098AC6BEDC38B0478A3099E03F4E55D46429DD365A359706557916137A2137316939F502F2BE428B8A9802EA9A0A0C4A960580D128EDC24E5B4A"
+    ],
+    "last_height": 0,
+    "last_round": 0,
+    "last_step": 0
+  },
+  "before": []
+}
